pyproject.toml
[tool.poetry]
name = "unit-na-tihom-tgbot"
version = "0.1.0"
description = ""
authors = ["modernpacifist <modernpacifist@gmail.com>"]
readme = "README.md"
packages = [{include = "unit_na_tihom_tgbot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.4"
logging = "^0.4.9.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
